Add webscale jenkins build (#4)
This adds a Jenkinsfile to build the libpam-webscale-authenticator package. If
a master build is being done, then a new -webscalen suffixed tag will be
created and the package will be published to the trusty and xenial repos with a
'main' component. For a branch build, a timestamp suffix is added to the most
recent tag and the publishing component is 'test'. Note that this means that it
is not really expected that more than one version will be under test at once,
since there is only a single version stream within that component.